+++ balticbrawler/src/components/FurtherInfo.tsx
@@ -28,12 +28,14 @@ function FurtherInfo() {
             header: "Parken",
             value: (
                 <>
-                    <Typography>"Das Parken ist für euch kostenlos"</Typography>
-                    <Typography>Ihr bekommt vor Ort ein Parkticket</Typography>
+                    <Typography>
+                        Das Parken ist für euch kostenlos, ihr bekommt vor Ort
+                        ein Parkticket.
+                    </Typography>
                     <Typography>
                         Parken könnt ihr auf allen Parkplätzen des Geländes, wir
-                        empfehlen euch allerdings die Parkplätze, die wir rot
-                        eingekreist haben, zu nutzen.
+                        empfehlen euch allerdings die rot eingekreisten
+                        Parkplätze zu nutzen.
                     </Typography>
                 </>
             ),
@@ -56,28 +58,32 @@ function FurtherInfo() {
             value: (
                 <>
                     <Typography maxWidth={500}>
-                        Es wird an beiden Tagen ein tolles Buffet geben. Partner
-                        und Familienangehörige dürfen gegen Aufpreis am Buffet
-                        teilnehmen. Kontaktiert uns dafür aber bitte rechtzeitig
-                        (bis 04.07.25), damit wir uns mit der Küche abstimmen
-                        können. In der Pause von 16:45-17:15Uhr (Tag 1)
-                        servieren wir euch frische Waffeln.
+                        Es wird an beiden Tagen ein tolles Buffet geben.
                     </Typography>
+                    <Typography>
+                        Partner und Familienangehörige dürfen gegen Aufpreis am
+                        Buffet teilnehmen. Kontaktiert uns dafür aber bitte
+                        rechtzeitig (bis 04.07.25), damit wir uns mit der Küche
+                        abstimmen können.
+                    </Typography>
+                    <Typography>
+                        In der Pause von 16:45-17:15Uhr (Tag 1) servieren wir
+                        euch frische Waffeln.
+                    </Typography>
+
+                    <Box justifySelf="flex-start" mt={1}>
+                        <Link
+                            justifySelf="center"
+                            href="/Speiseplan_BeachClashMajor_2025.pdf"
+                        >
+                            <img
+                                width="100%"
+                                src="/Speiseplan_BeachClashMajor_2025.png"
+                                style={{ maxWidth: 500 }}
+                            />
+                        </Link>
+                    </Box>
                 </>
-            ),
-            footer: (
-                <Box justifySelf="center">
-                    <Link
-                        justifySelf="center"
-                        href="/Speiseplan_BeachClashMajor_2025.pdf"
-                    >
-                        <img
-                            width="100%"
-                            src="/Speiseplan_BeachClashMajor_2025.png"
-                            style={{ maxWidth: 500 }}
-                        />
-                    </Link>
-                </Box>
             ),
         },
     ];
@@ -85,7 +91,7 @@ function FurtherInfo() {
     return (
         <>
             <ZoomableImage img={BeachClashMajor_2025_Map} />
-            <RulesPackList tableContent={roomAndFoodContent} />
+            <RulesPackList id="essen" tableContent={roomAndFoodContent} />
         </>
     );
 }