+++ balticbrawler/src/components/RulesPackList.tsx
@@ -3,29 +3,28 @@ import { ReactNode } from "react";
 import useIsMobile from "../hooks/useIsMobile";
 
 function RulesPackList(props: {
+    id?: string;
     tableContent: {
-        header: string;
+        header: ReactNode;
         value: ReactNode;
         footer?: ReactNode;
     }[];
 }) {
     const [isMobile] = useIsMobile();
     return (
-        <Box justifySelf="center">
-            {props.tableContent.map((x) => (
-                <Box>
-                    <Box
-                        display="flex"
-                        flexDirection={isMobile ? "column" : "row"}
-                        my={1}
-                        maxWidth={968}
-                    >
-                        <Typography width={200}>{x.header}:</Typography>
-                        <Typography flex={1} mx={isMobile ? 4 : 0}>
-                            {x.value}
-                        </Typography>
-                    </Box>
-                    {x.footer}
+        <Box id={props.id} justifySelf="left">
+            {props.tableContent.map((x, i) => (
+                <Box
+                    key={i}
+                    display="flex"
+                    flexDirection={isMobile ? "column" : "row"}
+                    my={1}
+                    maxWidth={900}
+                >
+                    <Typography width={200}>{x.header}:</Typography>
+                    <Typography flex={1} mx={isMobile ? 4 : 0}>
+                        {x.value}
+                    </Typography>
                 </Box>
             ))}
         </Box>