+++ balticbrawler/src/components/TournamentFaq.tsx
@@ -1,30 +1,8 @@
-import {
-    Box,
-    Table,
-    TableBody,
-    TableCell,
-    TableContainer,
-    TableHead,
-    TableRow,
-    Typography,
-} from "@mui/material";
+import { Box, Typography } from "@mui/material";
 import { ReactNode } from "react";
 import RulesPackList from "./RulesPackList";
 
 function TournamentFaq() {
-    const matrix = [
-        { diff: "0", winner: 10, loser: 10 },
-        { diff: "1-5", winner: 11, loser: 9 },
-        { diff: "6-10", winner: 12, loser: 8 },
-        { diff: "11-15", winner: 13, loser: 7 },
-        { diff: "16-20", winner: 14, loser: 6 },
-        { diff: "21-25", winner: 15, loser: 5 },
-        { diff: "26-30", winner: 16, loser: 4 },
-        { diff: "31-35", winner: 17, loser: 3 },
-        { diff: "36-40", winner: 18, loser: 2 },
-        { diff: "41-45", winner: 19, loser: 1 },
-        { diff: "46+", winner: 20, loser: 0 },
-    ];
     const tableContent: { header: string; value: ReactNode }[] = [
         {
             header: "Schachuhr",
@@ -99,11 +77,7 @@ function TournamentFaq() {
                         um nach Möglichkeit keine zusätzliche Zeit zu
                         verbrauchen!
                     </Typography>
-                    <Typography
-                        mt={2}
-                        fontWeight="bold"
-                        sx={{ textDecoration: "underline" }}
-                    >
+                    <Typography mt={2} fontWeight="bold">
                         Beide Spieler haben keine Zeit mehr (Spielzeit ist
                         vorbei):
                     </Typography>
@@ -152,10 +126,7 @@ function TournamentFaq() {
                         unterschiedlich gebaut sind.
                     </Typography>
 
-                    <Typography
-                        fontWeight="bold"
-                        sx={{ textDecoration: "underline" }}
-                    >
+                    <Typography fontWeight="bold">
                         Beispiel: Ihr spielt Chaos Knights und wollt, dass Alle
                         einen Havoc Launcher haben, obwohl einige mit Heavy
                         Stubbern ausgestattet sind. Dies muss bei der
@@ -165,36 +136,7 @@ function TournamentFaq() {
             ),
         },
     ];
-    return (
-        <>
-            <RulesPackList tableContent={tableContent} />
-            <Box mx={10} maxWidth={700} justifySelf="center" width="100%">
-                <Typography variant="h4" justifySelf="center" mt={4}>
-                    20:0 Matrix
-                </Typography>
-                <TableContainer>
-                    <Table size="small">
-                        <TableHead>
-                            <TableRow>
-                                <TableCell>Differenz</TableCell>
-                                <TableCell>Gewinner</TableCell>
-                                <TableCell>Verlierer</TableCell>
-                            </TableRow>
-                        </TableHead>
-                        <TableBody>
-                            {matrix.map((x) => (
-                                <TableRow>
-                                    <TableCell>{x.diff}</TableCell>
-                                    <TableCell>{x.winner}</TableCell>
-                                    <TableCell>{x.loser}</TableCell>
-                                </TableRow>
-                            ))}
-                        </TableBody>
-                    </Table>
-                </TableContainer>
-            </Box>
-        </>
-    );
+    return <RulesPackList tableContent={tableContent} />;
 }
 
 export default TournamentFaq;