+++ balticbrawler/src/hooks/atoms.ts
@@ -1,0 +1,10 @@
+import { atomWithHash } from "jotai-location";
+
+const zoomImageAtom = atomWithHash<string>("zoomImage", "", {
+                setHash: "default",
+                deserialize: (x) => x,
+                serialize: (x) => x,
+            });
+
+export default zoomImageAtom;
+